backend/requirements.txt
Flask==2.3.2
Flask_SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
Werkzeug==2.3.3
Flask-WTF==1.1.1
Flask-Login==0.6.2
